Seperated the Photos and Videos from Gallery
